requirements.txt
matplotlib==3.5.1
numpy==1.24.4
pandas==2.2.3
Pillow==11.0.0
scikit_image==0.24.0
scikit_learn==1.5.2
seaborn==0.13.2
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
transformers==4.25.1
ydata_profiling==4.10.0